Create NumberOfDaysInMonth
The program is supposed to take the name of a month in the text box, then return the correct number of days for that month. But no matter what I put into the box, it just gives the correct answer for January. I'm missing something but I've gone over and over this and I can't see what I'm doing wrong, though I'm sure it's something tiny and obvious.